epsilon added to user interface

+++ src/ConstMult/IntConstMultShiftAdd.hpp
@@ -19,7 +19,7 @@ namespace flopoco {
         static ostream nostream;
         int noOfPipelineStages;
 
-        IntConstMultShiftAdd(Operator* parentOp, Target* target,int wIn_, string pipelined_realization_str, bool pipelined_=true, bool syncInOut_=true, int syncEveryN_=1,bool syncMux_=true, double epsilon_=0);
+        IntConstMultShiftAdd(Operator* parentOp, Target* target,int wIn_, string pipelined_realization_str, bool pipelined_=true, bool syncInOut_=true, int syncEveryN_=1,bool syncMux_=true, int epsilon_=0);
 
         ~IntConstMultShiftAdd() {}
